requirements.txt
Django>=3.1.7
djangorestframework>=3.10.0
django-injector
pytest-django>=4.4.0
